requirements.txt
flask
supabase 
python-dotenv
